feat: Use one table function per TPCH table
This commit makes the change to have one table function for each TPCH
table we want to generate.

The idea behind this is that table functions are scoped to a single
table so we can have the granularity exposed to the user who wishes
to generate a single table and at the same time expose a top level
UDF that generates all the tables and builds the catalog implicitely
for the user.